change itakushacode type

+++ Models/GW1004/Response/ItakushaCode.cs
@@ -2,8 +2,8 @@
 {
     public class ItakushaCode
     {
-        public int sofuriItakushaCode { get; set; }
-        public int kyuyofurikomiItakushaCode { get; set; }
-        public int shoyofurikomiItakushaCode { get; set; }
+        public long sofuriItakushaCode { get; set; }
+        public long kyuyofurikomiItakushaCode { get; set; }
+        public long shoyofurikomiItakushaCode { get; set; }
     }
 }
